Refactor PasswordPrompt
